requirements.txt

flask==1.1.2
onnxruntime==1.8.1
numpy==1.21.2
Pillow==8.3.1
